Fixup tests and names
